Full SQL Translation txt File
Pushed a .zip file with all the .csv files used along with C++ code to format the .csv's. The SQLText.txt file is the file filled in with all of the appropriate SQL commands to be placed into our frontend database simulator for testing, along with Oracle at a later date.